webR: only include an image in result if it's non-empty
When you open the `svg()` device, it creates a file `Rplot001.svg`, but
hasn't written anything to it yet. If the plot is produced in a later
code block, then the code that looks for images will see the file before
it has anything in it.

There's no way that I can see of checking if a file is open, so this
just waits for an image file's contents to be non-empty.

+++ programming.js
@@ -664,8 +664,10 @@ Numbas.addExtension('programming', ['display', 'util', 'jme'], function(programm
                             const data = await webR.FS.readFile(path);
                             const blob = new Blob([data], {type: 'image/svg+xml'});
                             const text = await blob.text();
-                            images.push(text);
-                            await webR.FS.unlink(path);
+                            if(text) {
+                                images.push(text);
+                                await webR.FS.unlink(path);
+                            }
                         }
                     }
 
@@ -966,11 +968,11 @@ Numbas.addExtension('programming', ['display', 'util', 'jme'], function(programm
             }
             return new types.TPromise(programming.language_runners['webr'].load_webR().then(async (webR) => {
                 var promises = [];
-                files.forEach(filename => {
+                filenames.forEach(filename => {
                     promises.push(new Promise((resolve,reject) => {
                         setTimeout(async function() {
                             try {
-                                const buf = await webR.getFileData(filename);
+                                const buf = await webR.FS.readFile(filename);
                                 const extension = filename.match(/\.[^.]*$/)[0];
                                 const mime_type = mime_types[extension] || 'text/plain';
                                 var blob = new Blob([buf], {type: mime_type});
@@ -988,7 +990,7 @@ Numbas.addExtension('programming', ['display', 'util', 'jme'], function(programm
                 });
                 const results = await Promise.all(promises);
                 var o = {};
-                results.forEach((value,i) => { o[files[i]] = jme.wrapValue(value); });
+                results.forEach((value,i) => { o[filenames[i]] = jme.wrapValue(value); });
                 return {
                     r_files: new Numbas.jme.types.TDict(o)
                 }
